requirements.txt
pytesseract==0.3.10
numpy==1.26.4
Pillow==10.2.0
PyMuPDF==1.23.25
sentence-transformers==2.6.1
